[libc] Split 'stdio' module
