+++ libc/stdio/Makefile
@@ -2,7 +2,39 @@
 
 include $(TOPDIR)/libc/Makefile.inc
 
-SRCS = stdio.c printf.c scanf.c putchar.c
+SRCS = \
+	stdio.c \
+	__fopen.c \
+	__io_init.c \
+	__io_list.c \
+	fclose.c \
+	fdopen.c \
+	fflush.c \
+	fgetc.c \
+	fgets.c \
+	fopen.c \
+	fputc.c \
+	fputs.c \
+	fread.c \
+	freopen.c \
+	fseek.c \
+	ftell.c \
+	fwrite.c \
+	puts.c \
+	gets.c \
+	printf.c \
+	putchar.c \
+	rewind.c \
+	scanf.c \
+	setbuffer.c \
+	setvbuf.c \
+	stderr.c \
+	stdin.c \
+	stdout.c \
+	printf.c \
+	putchar.c \
+	ungetc.c \
+
 OBJS = $(SRCS:.c=.o)
 
 $(OBJS): $(SRCS)
