+++ libc/stdio/__io_list.c
@@ -1,0 +1,3 @@
+#include "_stdio.h"
+
+FILE *__IO_list = 0;		/* For fflush at exit */
