+++ libc/stdio/rewind.c
@@ -1,0 +1,10 @@
+#ifdef L_rewind
+#include <stdio.h>
+
+void
+rewind(FILE *fp)
+{
+   fseek(fp, (long)0, 0);
+   clearerr(fp);
+}
+#endif
